% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_api.R
\name{countrows}
\alias{countrows}
\title{Count rows in a table}
\usage{
countrows(api, table_id)
}
\arguments{
\item{api}{A 'gristapi::grist_api'-object}

\item{table_id}{Normalized table name (see \code{id} in \code{listtables} method). First character upper. OBLIGATORY}
}
\value{
A integer.
}
\description{
Returns the number of rows in a table
}
\examples{
# See number of rows
table_name <- paste0("Mtcars_", get_os())
api <- grist_api$new(
  server = 'https://grist.numerique.gouv.fr',
  api_key = Sys.getenv("GRIST_KEY"),
  doc_id = Sys.getenv("GRIST_DOC_TEST")
)
countrows(api, table_name)

}
\seealso{
Other function_gristapi: 
\code{\link{add_records}()},
\code{\link{delete_records}()},
\code{\link{fetch_table}()},
\code{\link{listcolumns}()},
\code{\link{listtables}()},
\code{\link{sync_table}()},
\code{\link{update_records}()}
}
\concept{function_gristapi}
