% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_api.R
\name{listcolumns}
\alias{listcolumns}
\title{List columns of a table}
\usage{
listcolumns(api, table_id, hidden = FALSE)
}
\arguments{
\item{api}{A 'gristapi::grist_api'-object}

\item{table_id}{Normalized table name (see \code{id} in \code{listtables} method). First character upper. OBLIGATORY}

\item{hidden}{Set to TRUE to include the hidden columns. Default FALSE}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} object.
}
\description{
Fetch the specs of columns in a target Grist document table
}
\examples{
# Working only if Env Var "GRIST_KEY" and "GRIST_DOC_TEST" are defined
# see the structure of a given table
table_name <- paste0("Mtcars_", get_os())
api <- grist_api$new(
  server = 'https://grist.numerique.gouv.fr',
  api_key = Sys.getenv("GRIST_KEY"),
  doc_id = Sys.getenv("GRIST_DOC_TEST")
)
listcolumns(api, table_id = table_name)

}
\seealso{
Other function_gristapi: 
\code{\link{add_records}()},
\code{\link{countrows}()},
\code{\link{delete_records}()},
\code{\link{fetch_table}()},
\code{\link{listtables}()},
\code{\link{sync_table}()},
\code{\link{update_records}()}
}
\concept{function_gristapi}
