% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_api.R
\name{listtables}
\alias{listtables}
\title{List tables}
\usage{
listtables(api)
}
\arguments{
\item{api}{A 'gristapi::grist_api'-object}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} object.
}
\description{
List tables in the target Grist document
}
\examples{
# List the tables in the document
api <- grist_api$new(
  server = 'https://grist.numerique.gouv.fr',
  api_key = Sys.getenv("GRIST_KEY"),
  doc_id = Sys.getenv("GRIST_DOC_TEST")
)
listtables(api)
}
\seealso{
Other function_gristapi: 
\code{\link{add_records}()},
\code{\link{countrows}()},
\code{\link{delete_records}()},
\code{\link{fetch_table}()},
\code{\link{listcolumns}()},
\code{\link{sync_table}()},
\code{\link{update_records}()}
}
\concept{function_gristapi}
